refactor: split up providers tests between providers and accounts packages
